+++ session_cache_redis.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"github.com/go-redis/redis/v8"
+	"log"
 	"time"
 )
 
@@ -11,13 +12,19 @@ import (
 type SessionCacheRedis struct {
 	maxAge time.Duration
 	rds    *redis.Client
+	logger *log.Logger
 }
 
 // NewSessionCacheRedis creates a new cache for given maximum age and redis connection.
-func NewSessionCacheRedis(maxAge time.Duration, redisOptions *redis.Options) *SessionCacheRedis {
+func NewSessionCacheRedis(maxAge time.Duration, log *log.Logger, redisOptions *redis.Options) *SessionCacheRedis {
+	if log == nil {
+		log = logger
+	}
+
 	return &SessionCacheRedis{
 		maxAge: maxAge,
 		rds:    redis.NewClient(redisOptions),
+		logger: log,
 	}
 }
 
@@ -26,12 +33,17 @@ func (cache *SessionCacheRedis) Get(clientID, fingerprint uint64, _ time.Time) *
 	r, err := cache.rds.Get(context.Background(), getSessionKey(clientID, fingerprint)).Result()
 
 	if err != nil {
+		if err != redis.Nil {
+			cache.logger.Printf("error reading session from cache: %s", err)
+		}
+
 		return nil
 	}
 
 	var session Session
 
 	if err := json.Unmarshal([]byte(r), &session); err != nil {
+		cache.logger.Printf("error unmarshalling session from cache: %s", err)
 		return nil
 	}
 
@@ -44,6 +56,8 @@ func (cache *SessionCacheRedis) Put(clientID, fingerprint uint64, session *Sessi
 
 	if err == nil {
 		cache.rds.SetEX(context.Background(), getSessionKey(clientID, fingerprint), v, cache.maxAge)
+	} else {
+		cache.logger.Printf("error storing session in cache: %s", err)
 	}
 }